Convert Eclipse Project to Maven
Project structure was not 'Maven standard', so  the `main` and `test` source folders have been moved into sub-folders so that they confirm to the recommended layout (as detailed at https://maven.apache.org/guides/introduction/introduction-to-the-standard-directory-layout.html)

Confirmed that Maven build and test cycle completes without error and produces a JAR at the end.

In Eclipse, simply right-click the `pom.xml` file and (from the context menu which appears) select `Run as...` ► `Maven build`. If prompted, set the "Goals" option to `clean verify` before running the build.